add generic 'convert_species'
By default
```
convert_species(some_data_frame, homologues)
```

returns a data-frame with the same number of columns as
`some_data_frame`, but where the homologues data-frame is used to
convert the IDs in the first column (of some_data_frame) to the
corresponding ID from the second column of homologues. The column name
for the returned data-frame is taken from the second column of
homologues.

User can specify
- where in `some_data_frame` the ID column is;
- what the replacement column name should be.

Also
- added shorthand for making tibbles (used in the tests for
`convert_species`)